+++ SpiceQL/include/SpiceQL/io.h
@@ -158,12 +158,10 @@ namespace SpiceQL {
       *
       * @param fileName path to file to write the segment to
       * @param quats nx4 vector of orientations as quaternions
-      * @param times nx1 vector of times matching the number of quats
+      * @param times nx1 vector of encoded SCLK ticks
       * @param bodyCode NAIF body code identifying the orientations belong to
       * @param referenceFrame NAIF string for the reference frame the orientations are in
       * @param segmentId Some ID to give the segment
-      * @param sclk path to S clock kernal to convert to and from ephemeris time
-      * @param lsk path to leap second kernal 
       * @param angularVelocity optional, nx3 array of angular velocities
       * @param comment optional, comment to be associated with the segment
       */
@@ -173,8 +171,6 @@ namespace SpiceQL {
                  int bodyCode,
                  std::string referenceFrame,
                  std::string segmentId,
-                 std::string sclk,
-                 std::string lsk,
                  std::vector<std::vector<double>> angularVelocity = {},
                  std::string comment = "");
 
@@ -183,15 +179,12 @@ namespace SpiceQL {
      * @brief Write CK segments to a file
      *
      * Given orientations, angular velocities, times, target and reference frames, write data as a segment in a CK kernel.
+     * The segments' times must already be encoded SCLK ticks.
      *
      * @param fileName path to file to write the segment to
-     * @param sclk path to SCLK kernel matching the segments' frame code
-     * @param lsk path to LSK kernel
      * @param segments spkSegments to be writte
      */
   void writeCk(std::string fileName,
-               std::string sclk,
-               std::string lsk,
                std::vector<CkSegment> segments);
 
   void writeComment(std::string fileName,